Convert skills providers to the Skills plugin
Sixth and final Phase 1 conversion under FMCP-20. Moves
fastmcp.server.providers.skills/*.py into
fastmcp.server.plugins.skills/*.py and wraps the whole family in a
single Skills(Plugin[SkillsConfig]) plugin. SkillsConfig collapses the
nine pre-plugin classes (SkillProvider, SkillsDirectoryProvider, seven
vendor subclasses) into one discriminated config with
`path` / `directory` / `vendor` fields.

Backcompat shim at the old path: package __init__ silently re-exports
the underlying classes, leaf submodules emit FastMCPDeprecationWarning.
The seven vendor subclasses remain importable for direct use.

Closes FMCP-30.
